Add support for unsorted CSV files
Two new options are provided -csv1.sort and -csv2.sort, which if either
are set, will sort the corresponding data by the -key columns before
diffing with the other source.

This uses SQLite behind the scenes which requires CGO. This release drops
Windows binaries for the time being until an alternate implementation is
considered.

Signed-off-by: Byron Ruth <b@devel.io>