Drop SQLite dependency for simple sort implementation
It was assumed that the SQLite implementation would be faster due to
inherent optimizations, but this was completely false.

For example, a ~166 MB CSV took ~1m13s with the SQLite implementation
and now takes ~12s with pure Go solution. It also takes a bit ~10% less
memory.

The other advantage is C Go is no longer needed so Windows builds are
available again.

Signed-off-by: Byron Ruth <b@devel.io>