Add -snapshot option for snapshotting table state
The purpose of this is to support initializing the state of a stream by
snapshotting the initial state of the table and as well as creating
snapshots of the table on regular intervals. This reduces the overhead
for consumers to produce the current state by reading the stream. A
snapshot enables a consumer to start at the last known snapshot and read
from there rather than the beginning of the stream.

Signed-off-by: Byron Ruth <b@devel.io>